Refactor contentManager file to be more organized. Updated client side typing
